requirements-common.txt
# Common requirements shared across ALL Face Recognition services
# Only dependencies that are used by ALL services should be here

# Core numerical stack
numpy==1.23.5
PyYAML==6.0.2

# Messaging and storage clients shared across services
redis==5.0.7
pika==1.3.2
aio-pika==9.4.1
minio==7.2.7
pynvml==11.5.0

# Computer vision utilities leveraged in common utilities
opencv-python-headless==4.10.0.84
Pillow==10.4.0
orjson==3.10.6
